DL-17282 | Change Identifier type from Int to Long (#576)

+++ app/models/pla/response/AmendProtectionResponse.scala
@@ -20,7 +20,7 @@ import models.pla.{AmendProtectionLifetimeAllowanceType, AmendProtectionResponse
 import play.api.libs.json.{Format, Json}
 
 case class AmendProtectionResponse(
-    lifetimeAllowanceIdentifier: Int,
+    lifetimeAllowanceIdentifier: Long,
     lifetimeAllowanceSequenceNumber: Int,
     lifetimeAllowanceType: AmendProtectionLifetimeAllowanceType,
     certificateDate: Option[String],
